✨ feat: 헤더에 전역 상태 관리 반영, test component파일 삭제
