+++ app/_components/header/_component/header-login.tsx
@@ -1,12 +1,10 @@
 import Image from "next/image";
 import Link from "next/link";
-
 interface HeaderLoginProps {
-  onClick: () => void;
-  type: "employer" | "employee";
+  type: "employer" | "employee" | null;
+  logout: () => void;
 }
-
-export default function HeaderLogin({ onClick, type }: HeaderLoginProps) {
+export default function HeaderLogin({ type, logout }: HeaderLoginProps) {
   return (
     <>
       <section>
@@ -16,7 +14,7 @@ export default function HeaderLogin({ onClick, type }: HeaderLoginProps) {
           <Link href="/profile-detail">내 프로필</Link>
         )}
       </section>
-      <button onClick={onClick}>로그아웃</button>
+      <button onClick={logout}>로그아웃</button>
       <section className="relative h-5 w-5 md:h-6 md:w-6">
         <Image
           src="/icons/noti-none.png"
